move courseTable API to file
